totalCount修改为long型

+++ common-utils/src/main/java/com/caishi/capricorn/common/pager/CommonPager.java
@@ -8,16 +8,16 @@ import java.util.List;
 public class CommonPager<T> {
 
     /** total count */
-    private Integer totalCount = 0;
+    private long totalCount = 0;
 
     /** data */
     private List<T> data = null;
 
-    public Integer getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public void setTotalCount(Integer totalCount) {
+    public void setTotalCount(long totalCount) {
         this.totalCount = totalCount;
     }
